Break out AtracTrack into its own file. Add a little atrac analysis tool to ImDebugger (for future use)
