+++ Core/Core.vcxproj.filters
@@ -1342,6 +1342,9 @@
     <ClCompile Include="HLE\sceAac.cpp">
       <Filter>HLE\Libraries</Filter>
     </ClCompile>
+    <ClCompile Include="Util\AtracTrack.cpp">
+      <Filter>Util</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -2165,6 +2168,9 @@
     </ClInclude>
     <ClInclude Include="HLE\ErrorCodes.h">
       <Filter>HLE</Filter>
+    </ClInclude>
+    <ClInclude Include="Util\AtracTrack.h">
+      <Filter>Util</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
