+++ UI/ImDebugger/ImDebugger.h
@@ -150,6 +150,7 @@ struct ImConfig {
 	bool sasAudioOpen;
 	bool logConfigOpen;
 	bool utilityModulesOpen;
+	bool atracToolOpen;
 	bool memViewOpen[4];
 
 	// HLE explorer settings
@@ -176,11 +177,22 @@ struct ImConfig {
 
 	bool sasShowAllVoices = false;
 
+
 	// We use a separate ini file from the main PPSSPP config.
 	void LoadConfig(const Path &iniFile);
 	void SaveConfig(const Path &iniFile);
 
 	void SyncConfig(IniFile *ini, bool save);
+};
+
+struct Track;
+class ImAtracToolWindow {
+public:
+	void Draw(ImConfig &cfg);
+
+	char atracPath_[1024]{};
+	std::unique_ptr<Track> track_;
+	std::string error_;
 };
 
 enum class ImCmd {
@@ -233,6 +245,7 @@ private:
 	ImStructViewer structViewer_;
 	ImGePixelViewerWindow pixelViewer_;
 	ImMemDumpWindow memDumpWindow_;
+	ImAtracToolWindow atracToolWindow_;
 
 	ImSnapshotState newSnapshot_;
 	ImSnapshotState snapshot_;
